new files

+++ src/main/java/com/eventum/backend/StudentServiceImpl.java
@@ -1,21 +1,21 @@
 package com.eventum.backend;
 
-import com.eventum.backend.EntityExampleListingParamsDto;
-import org.springframework.beans.factory.annotation.Autowired;
-import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
+
 @Service
-public class StudentServiceImpl implements StudentService {
+public class StudentServiceImpl implements StudentService{
 
   @Autowired
   private StudentRepository studentRepository;
 
   @Override
   @Transactional
-  public Iterable<Student> findAll(EntityExampleListingParamsDto params) {
+  public List<StudentDTO> findAll(StudentListingParamsDto params) {
     // String searchQuery = StringUtils.escapeSQLSearchQuery(params.getSearchQuery());
     // DynamicQueryBuilder<Student> dynamicQuery = dynamicQueryService.build(Student.class);
     // dynamicQuery.or(new FilterDto<>(NAME, Boolean.FALSE, LIKE_CHAR + searchQuery + LIKE_CHAR));
@@ -23,32 +23,44 @@ public class StudentServiceImpl implements StudentService {
     // dynamicQuery.paginate(params.getPaginationNumber() * params.getPaginationRange(), params.getPaginationRange());
     // List<Student> resultList = dynamicQuery.execute(entityManager);
     // return TableListingUtils.generateResultMap(mapper.mapEntityExamples(resultList), dynamicQuery.getResultSize(entityManager));
-    return  studentRepository.findAll();
+    return ((List<Student>)studentRepository.findAll())
+    .stream()
+    .map(x -> Mapper.mapToDTO(x))
+    .collect(Collectors.toList());
   }
 
   @Override
   @Transactional
-  public Student findById(Integer id) {
+  public StudentDTO findById(Long id) {
     Optional<Student> opt = studentRepository.findById(id);
-    return  opt.isPresent()? opt.get(): null;
+    return opt.isPresent()?Mapper.mapToDTO(opt.get()): null;
   }
 
   @Override
   @Transactional
-  public void save(Student entity) {
+  public Boolean save(StudentDTO dto) {
+    Student entity = Mapper.mapFromDTO(dto);
     studentRepository.save(entity);
+    return true;
   }
 
   @Override
   @Transactional
-  public void update(Student entity) {
+  public Boolean update(StudentDTO dto) {
+    Student entity = Mapper.mapFromDTO(dto);
     studentRepository.save(entity);
+    return true;
   }
 
   @Override
   @Transactional
-  public void deleteById(Integer id) {
-    studentRepository.deleteById(id);
+  public Boolean deleteById(Long id) {
+    Optional<Student> res = studentRepository.findById(id);
+    if(res.isPresent()){
+      studentRepository.deleteById(id);
+      return true;
+    }
+    return false;
   }
 
 }
